+++ src/main/java/endpoint/CreateUserHttpRequestHandle.java
@@ -1,0 +1,14 @@
+package endpoint;
+
+public class CreateUserHttpRequestHandle extends HttpRequestEndpoint {
+    private static final String ENDPOINT = "/user/create";
+
+    public CreateUserHttpRequestHandle() {
+        super(ENDPOINT);
+    }
+
+    @Override
+    public void handle() {
+
+    }
+}
