+++ src/main/java/endpoint/HttpRequestHandle.java
@@ -1,0 +1,5 @@
+package endpoint;
+
+public interface HttpRequestHandle {
+    void handle();
+}
