+++ src/main/java/endpoint/NoneEndPoint.java
@@ -1,0 +1,12 @@
+package endpoint;
+
+public class NoneEndPoint extends HttpRequestEndpoint {
+
+    public NoneEndPoint() {
+        super("");
+    }
+
+    @Override
+    public void handle() {
+    }
+}
