+++ src/main/java/webserver/RequestHandler.java
@@ -1,10 +1,17 @@
 package webserver;
 
+import endpoint.RequestEndpointRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import utils.FileIoUtils;
 import utils.IOUtils;
+import webserver.http.HttpRequestMessage;
+import webserver.http.request.HttpRequestMessageParser;
 
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.DataOutputStream;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.net.Socket;
 import java.util.List;
 
@@ -27,34 +34,17 @@ public class RequestHandler implements Runnable {
                 BufferedReader br = new BufferedReader(new InputStreamReader(connection.getInputStream(), UTF_8));
                 OutputStream out = connection.getOutputStream()
         ) {
-
             List<String> httpRequestLines = IOUtils.readLines(br);
+            HttpRequestMessage httpRequestMessage = HttpRequestMessageParser.parse(httpRequestLines);
+
+            String httpPath = httpRequestMessage.httpPath();
+            RequestEndpointRegistry.getEndpoint(httpPath);
 
             DataOutputStream dos = new DataOutputStream(out);
-            byte[] body = "Hello World".getBytes();
-            response200Header(dos, body.length);
-            responseBody(dos, body);
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-    }
-
-    private void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-    }
-
-    private void responseBody(DataOutputStream dos, byte[] body) {
-        try {
-            dos.write(body, 0, body.length);
-            dos.flush();
-        } catch (IOException e) {
+            byte[] body = FileIoUtils.loadFileFromClasspath("./templates" + httpRequestMessage.httpPath());
+            ResponseHandler.response200Header(dos, body.length);
+            ResponseHandler.responseBody(dos, body);
+        } catch (Exception e) {
             logger.error(e.getMessage());
         }
     }
