+++ src/main/java/webserver/WebApplicationServer.java
@@ -1,10 +1,11 @@
 package webserver;
 
-import java.net.ServerSocket;
-import java.net.Socket;
-
+import endpoint.RequestEndpointRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.net.ServerSocket;
+import java.net.Socket;
 
 public class WebApplicationServer {
     private static final Logger logger = LoggerFactory.getLogger(WebApplicationServer.class);
@@ -21,6 +22,7 @@ public class WebApplicationServer {
         // 서버소켓을 생성한다. 웹서버는 기본적으로 8080번 포트를 사용한다.
         try (ServerSocket listenSocket = new ServerSocket(port)) {
             logger.info("Web Application Server started {} port.", port);
+            RequestEndpointRegistry.initializeEndpointRegistry();
 
             // 클라이언트가 연결될때까지 대기한다.
             Socket connection;
