+++ src/main/java/webserver/http/request/requestline/HttpMethod.java
@@ -1,0 +1,14 @@
+package webserver.http.request.requestline;
+
+import java.util.Arrays;
+
+public enum HttpMethod {
+    GET, POST;
+
+    public static HttpMethod of(String methodValue) {
+        return Arrays.stream(values())
+                .filter(httpMethod -> methodValue.equals(httpMethod.name()))
+                .findAny()
+                .orElseThrow(() -> new IllegalArgumentException(String.format("유효하지 않은 HTTP method입니다. HttpMethod: [%s]", methodValue)));
+    }
+}
