+++ src/main/java/webserver/http/response/HttpResponseBody.java
@@ -1,0 +1,9 @@
+package webserver.http.response;
+
+public class HttpResponseBody {
+    private byte[] bodyBytes;
+
+    public HttpResponseBody(byte[] bodyBytes) {
+        this.bodyBytes = bodyBytes;
+    }
+}
