+++ src/main/java/webserver/http/response/HttpResponseMessage.java
@@ -1,0 +1,9 @@
+package webserver.http.response;
+
+public class HttpResponseMessage {
+    public static final String RESPONSE_END_OF_LINE_MARKER = "\r\n";
+
+    private HttpResponseBody httpResponseBody;
+    private HttpResponseHeaders httpResponseHeaders;
+
+}
