+++ src/main/java/webserver/http/response/HttpResponseStatusLine.java
@@ -1,0 +1,8 @@
+package webserver.http.response;
+
+import webserver.http.HttpProtocolSchema;
+
+public class HttpResponseStatusLine {
+    private HttpProtocolSchema httpProtocolSchema;
+    private HttpStatus httpStatus;
+}
