--- src/main/java/webserver/httprequest/requestline/HttpVersion.java
@@ -1,20 +1,0 @@
-package webserver.httprequest.requestline;
-
-import java.util.Arrays;
-
-public enum HttpVersion {
-    ONE("1.0"), ONE_POINT_ONE("1.1"),
-    ;
-    private final String httpVersionValue;
-
-    HttpVersion(String httpVersionValue) {
-        this.httpVersionValue = httpVersionValue;
-    }
-
-    public static HttpVersion of(String httpVersionValue) {
-        return Arrays.stream(values())
-                .filter(httpVersion -> httpVersionValue.equals(httpVersion.httpVersionValue))
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException(String.format("잘못된 HTTP version httpVersionValue:[%s]", httpVersionValue)));
-    }
-}
