--- src/test/java/webserver/httprequest/header/HttpRequestHeadersTest.java
@@ -1,22 +1,0 @@
-package webserver.httprequest.header;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class HttpRequestHeadersTest {
-    @Test
-    void create_and_get() {
-        HttpRequestHeaders httpRequestHeaders = new HttpRequestHeaders(List.of(
-                "Host: localhost:8080",
-                "Connection: keep-alive",
-                "Accept: */*"
-        ));
-
-        assertThat(httpRequestHeaders.get("Host")).isEqualTo("localhost:8080");
-        assertThat(httpRequestHeaders.get("Connection")).isEqualTo("keep-alive");
-        assertThat(httpRequestHeaders.get("Accept")).isEqualTo("*/*");
-    }
-}
